requirements.txt
click==8.0.0
colorama==0.4.4
Flask==2.0.0
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
itsdangerous==2.0.0
Jinja2==3.0.0
MarkupSafe==2.0.0
mysql-connector-python==8.0.25
mysqlclient==2.0.3
numpy==1.20.3
pandas==1.2.4
protobuf==3.17.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.15
Werkzeug==2.0.0
